+++ macros/date_to_string.sql
@@ -1,7 +1,7 @@
 {# ------------------------------------------------------------------------------
 @Author:        F. Paul Spitzner
 @Created:       2025-07-17 13:19:58
-@Last Modified: 2025-08-11 13:33:51
+@Last Modified: 2025-08-13 17:45:41
 ------------------------------------------------------------------------------ #}
 
 {# macrodocs
@@ -37,6 +37,9 @@ lf_utils.date_to_string('my_date', 104)
 -- Get month and year in text format
 lf_utils.date_to_string('my_date', 'MMMM yyyy')
 ```
+
+# Notes
+- Consider using `lf_utils.date_strftime` for more flexible date formatting.
 
 endmacrodocs #}